Rename unit tests project. Clean up unused files. Move CodeGenApp to SharpGen.Interactive since it is only used by SharpGen.Interactive now.
