--- SharpGen.E2ETests/GlobalSuppressions.cs
@@ -1,8 +1,0 @@
-﻿
-// This file is used by Code Analysis to maintain SuppressMessage 
-// attributes that are applied to this project.
-// Project-level suppressions either have no target or are given 
-// a specific target and scoped to a namespace, type, member, etc.
-
-[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "xUnit1004:Test methods should not be skipped", Justification = "<Pending>", Scope = "member", Target = "~M:SharpGen.E2ETests.EnumTests.CppEnumWithDifferentUnderlyingTypeMapsToCSharpEnum")]
-
