+++ SharpGen.UnitTests/Mapping/Function.cs
@@ -8,7 +8,7 @@ using System.Text;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace SharpGen.E2ETests.Mapping
+namespace SharpGen.UnitTests.Mapping
 {
     public class Function : MappingTestBase
     {
