+++ SharpGen.UnitTests/Parsing/Enum.cs
@@ -6,7 +6,7 @@ using System.Text;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace SharpGen.E2ETests.Parsing
+namespace SharpGen.UnitTests.Parsing
 {
     public class Enum : ParsingTestBase
     {
