+++ SharpGen/SharpGen.csproj
@@ -5,7 +5,6 @@
     <Description>Library for generating efficient C# code from C++ and COM headers.</Description>
   </PropertyGroup>
   <ItemGroup>
-    <EmbeddedResource Include="**/*.tt" />
     <PackageReference Include="HtmlAgilityPack" Version="1.5.0-beta9" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="2.1.0" />
     <PackageReference Include="Microsoft.Win32.Registry" Version="4.3.0" />
@@ -14,7 +13,6 @@
     <PackageReference Include="System.IO.Compression" Version="4.3.0" />
     <PackageReference Include="System.Net.Requests" Version="4.3.0" />
     <PackageReference Include="System.Runtime.Extensions" Version="4.3.0" />
-    <PackageReference Include="System.Runtime.Loader" Version="4.3.0" Condition="'$(TargetFramework)' == 'netstandard1.5'" />
     <PackageReference Include="System.ServiceModel.Duplex" Version="4.3.0" />
     <PackageReference Include="System.ServiceModel.Http" Version="4.3.0" />
     <PackageReference Include="System.ServiceModel.NetTcp" Version="4.3.0" />
